Changed the name of the package and project inside all the gradle files to LaunchCash instead of fitnesstracker
