+++ LaunchCash-Backend_Database/build.gradle
@@ -4,7 +4,7 @@ plugins {
 	id 'io.spring.dependency-management' version '1.1.5'
 }
 
-group = 'com.hodges'
+group = 'com.CraftyCoders'
 version = '0.0.1-SNAPSHOT'
 
 java {
