+++ LaunchCash-Backend_Database/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'fitnessTracker1'
+rootProject.name = 'LaunchCash'
