+++ LaunchCash-Backend_Database/src/main/java/com/CraftyCoders/LaunchCash/LaunchCashApplication.java
@@ -1,0 +1,13 @@
+package com.CraftyCoders.LaunchCash;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LaunchCashApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(LaunchCashApplication.class, args);
+	}
+
+}
