+++ LaunchCash-Backend_Database/src/main/java/com/CraftyCoders/LaunchCash/controllers/HomeController.java
@@ -1,0 +1,15 @@
+package com.CraftyCoders.LaunchCash.controllers;
+
+import jdk.jfr.Category;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+
+    @GetMapping()
+    public String home(){
+        return "home";
+    }
+
+}
