+++ LaunchCash-Backend_Database/src/main/java/com/CraftyCoders/LaunchCash/models/Tag.java
@@ -1,0 +1,47 @@
+package com.CraftyCoders.LaunchCash.models;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToMany;
+
+import java.util.HashSet;
+import java.util.Set;
+
+@Entity
+public class Tag {
+    @Id
+    @GeneratedValue
+    private Long id;
+    private String name;
+
+    @ManyToMany(mappedBy = "tags")
+    private Set<Activity> activities = new HashSet<>();
+
+    public Tag() {
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public Set<Activity> getActivities() {
+        return activities;
+    }
+
+    public void setActivities(Set<Activity> activities) {
+        this.activities = activities;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
