+++ LaunchCash-Backend_Database/src/main/java/com/CraftyCoders/LaunchCash/models/dto/RegisterFormDTO.java
@@ -1,0 +1,16 @@
+package com.CraftyCoders.LaunchCash.models.dto;
+
+public class RegisterFormDTO extends LoginFormDTO {
+
+    private String verifyPassword;
+
+    public String getVerifyPassword() {
+        return verifyPassword;
+    }
+
+    public void setVerifyPassword(String verifyPassword) {
+        this.verifyPassword = verifyPassword;
+    }
+
+
+}
