+++ LaunchCash-Backend_Database/src/main/java/com/CraftyCoders/LaunchCash/repositories/UserRepository.java
@@ -1,0 +1,10 @@
+package com.CraftyCoders.LaunchCash.repositories;
+
+import com.CraftyCoders.LaunchCash.models.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User,Long> {
+    User findByUsername(String username);
+}
