+++ LaunchCash-Backend_Database/src/main/java/com/CraftyCoders/LaunchCash/security/AuthenticationFilter.java
@@ -1,0 +1,59 @@
+package com.CraftyCoders.LaunchCash.security;
+
+import com.CraftyCoders.LaunchCash.repositories.UserRepository;
+import com.CraftyCoders.LaunchCash.controllers.AuthenticationController;
+import com.CraftyCoders.LaunchCash.models.User;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.servlet.HandlerInterceptor;
+
+import java.io.IOException;
+import java.util.Arrays;
+import java.util.List;
+
+public class AuthenticationFilter implements HandlerInterceptor {
+
+    @Autowired
+    UserRepository userRepository;
+
+    @Autowired
+    AuthenticationController authenticationController;
+
+    private static final List<String> whitelist = Arrays.asList("/login", "/register", "/logout", "/css");
+
+    private static boolean isWhitelisted(String path) {
+        for (String pathRoot : whitelist) {
+            if (path.startsWith(pathRoot)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    @Override
+    public boolean preHandle(HttpServletRequest request,
+                             HttpServletResponse response,
+                             Object handler) throws IOException {
+
+        // Don't require sign-in for whitelisted pages
+        if (isWhitelisted(request.getRequestURI())) {
+            // returning true indicates that the request may proceed
+            return true;
+        }
+
+        HttpSession session = request.getSession();
+        User user = authenticationController.getUserFromSession(session);
+
+        // The user is logged in
+        if (user != null) {
+            return true;
+        }
+
+        // The user is NOT logged in
+        response.sendRedirect("/login");
+        return false;
+    }
+
+}
