+++ LaunchCash-Backend_Database/src/main/java/com/CraftyCoders/LaunchCash/security/WebApplicationConfig.java
@@ -1,0 +1,24 @@
+package com.CraftyCoders.LaunchCash.security;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.HandlerInterceptor;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebApplicationConfig implements WebMvcConfigurer {
+
+    // Create spring-managed object to allow the app to access our filter
+    @Bean
+    public HandlerInterceptor authenticationFilter() {
+        return new AuthenticationFilter();
+    }
+
+    // Register the filter with the Spring container
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor( authenticationFilter() );
+    }
+
+}
