--- LaunchCash-Backend_Database/src/main/java/com/hodges/LaunchCash/LaunchCashApplication.java
@@ -1,13 +1,0 @@
-package com.hodges.LaunchCash;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LaunchCashApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(LaunchCashApplication.class, args);
-	}
-
-}
