--- LaunchCash-Backend_Database/src/main/java/com/hodges/LaunchCash/models/User.java
@@ -1,67 +1,0 @@
-package com.hodges.LaunchCash.models;
-
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotEmpty;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-
-@Entity
-public class User {
-
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @NotEmpty
-    private String username;
-
-    @NotEmpty
-    private String hashedPassword;
-
-    private static final BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
-
-    @OneToOne(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
-    private Profile profile;
-
-    public User() {
-    }
-    public User(String username, String hashedPassword){
-        this.username=username;
-        this.hashedPassword=encoder.encode(hashedPassword);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-
-    public Profile getProfile() {
-        return profile;
-    }
-
-    public void setProfile(Profile profile) {
-        this.profile = profile;
-    }
-
-    public String getHashedPassword() {
-        return hashedPassword;
-    }
-
-    public void setHashedPassword(String hashedPassword) {
-        this.hashedPassword = encoder.encode(hashedPassword);
-    }
-    public boolean isMatchingPassword(String password) {
-        return encoder.matches(password, hashedPassword);
-    }
-}
