--- LaunchCash-Backend_Database/src/main/java/com/hodges/LaunchCash/models/dto/RegisterFormDTO.java
@@ -1,16 +1,0 @@
-package com.hodges.LaunchCash.models.dto;
-
-public class RegisterFormDTO extends LoginFormDTO {
-
-    private String verifyPassword;
-
-    public String getVerifyPassword() {
-        return verifyPassword;
-    }
-
-    public void setVerifyPassword(String verifyPassword) {
-        this.verifyPassword = verifyPassword;
-    }
-
-
-}
