--- LaunchCash-Backend_Database/src/main/java/com/hodges/LaunchCash/repositories/ActivityRepository.java
@@ -1,13 +1,0 @@
-package com.hodges.LaunchCash.repositories;
-
-import com.hodges.LaunchCash.models.Activity;
-import com.hodges.LaunchCash.models.Tag;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface ActivityRepository extends JpaRepository<Activity,Long> {
-    List<Activity> findByTagsContains(Tag tag);
-}
