--- LaunchCash-Backend_Database/src/main/java/com/hodges/LaunchCash/repositories/TagRepository.java
@@ -1,9 +1,0 @@
-package com.hodges.LaunchCash.repositories;
-
-import com.hodges.LaunchCash.models.Tag;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TagRepository extends JpaRepository<Tag,Long> {
-}
