--- LaunchCash-Backend_Database/src/main/java/com/hodges/LaunchCash/repositories/TransactionRepository.java
@@ -1,10 +1,0 @@
-package com.hodges.LaunchCash.repositories;
-
-import com.hodges.LaunchCash.models.Transaction;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TransactionRepository extends JpaRepository <Transaction, Long> {
-
-}
