--- LaunchCash-Backend_Database/src/main/java/com/hodges/LaunchCash/repositories/UserRepository.java
@@ -1,10 +1,0 @@
-package com.hodges.LaunchCash.repositories;
-
-import com.hodges.LaunchCash.models.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<User,Long> {
-    User findByUsername(String username);
-}
