--- LaunchCash-Backend_Database/src/main/java/com/hodges/LaunchCash/repositories/userRepo.java
@@ -1,7 +1,0 @@
-package com.hodges.LaunchCash.repositories;
-
-import com.hodges.LaunchCash.models.dto.user;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface userRepo extends JpaRepository<user,Long> {
-}
