+++ LaunchCash-Backend_Database/src/test/java/com/CraftyCoders/LaunchCash/LaunchCashApplicationTests.java
@@ -1,0 +1,13 @@
+package com.CraftyCoders.LaunchCash;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class LaunchCashApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
