--- LaunchCash-Backend_Database/src/test/java/com/hodges/LaunchCash/LaunchCashApplicationTests.java
@@ -1,13 +1,0 @@
-package com.hodges.LaunchCash;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class LaunchCashApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
